unify codel params-clamp target to set values-do not adjust to apparent mtu.
We have a long standing "knowledge" that when target > than what the mtu
can serve at the available bandwidth...

		"bad things" happen to codel'd packets.

What we don't have, is extensive characterization of what that actually
looks like on a variety of tests.

So let's go get that data. THEN we can start talking about
mtu * 1.5 or any other value whatsoever, or gasp! even an
algorithm to calculate a smarter target based on the number of flows.