Get rid of configurable targets quantums and rtt
In my world, all the heavy flows are less than 40ms away.

100ms as an outside guess has been used since time immemorial
as a proxy for cross USA RTTs. It's untrue to start with (85ms),
and most of my heavy hitting flows are in the 40-50ms range.

as for a 6ms target, well, that dovetails with CMTS delays.

I reserve the right to change these.

Quantum is essentially a constant in the first place also.

This commit also kills GSO splitting. We're desparate
for CPU, if we get a burst like that, let something
further downstream split it apart.